[feat] 로그인 페이지 추가 및 유튜브 등록 수정 삭제 권한 추가
##### 제목은 최대 50 글자까지만 입력 ############## ->
- 로그인 페이지 추가
- 유튜브 등록 수정 삭제 권한 추가
- 미로그인시 허용권한 최소화

# 본문은 위에 작성
######## 본문은 한 줄에 최대 72 글자까지만 입력 ########################### -> |

# 꼬릿말은 아래에 작성: ex) #이슈 번호

# --- COMMIT END ---
#   <타입> 리스트
#   feat     : 기능 (새로운 기능)
#   fix      : 버그 (버그 수정)
#   refactor : 리팩토링
#   style    : 스타일 (코드 형식, 세미콜론 추가: 비즈니스 로직에 변경 없음)
#   docs     : 문서 (문서 추가, 수정, 삭제)
#   test     : 테스트 (테스트 코드 추가, 수정, 삭제: 비즈니스 로직에 변경 없음)
#   chore    : 기타 변경사항 (빌드 스크립트 수정 등)
#   lec  : 강의
# ------------------
#     제목 첫 글자를 대문자로
#     제목은 명령문으로
#     제목 끝에 마침표(.) 금지
#     제목과 본문을 한 줄 띄워 분리하기
#     본문은 "어떻게" 보다 "무엇을", "왜"를 설명한다.
#     본문에 여러줄의 메시지를 작성할 땐 "-"로 구분
# ------------------

+++ src/main/java/TwitchToYoutube/timeline/controller/YoutubeController.java
@@ -3,7 +3,8 @@ package TwitchToYoutube.timeline.controller;
 import TwitchToYoutube.timeline.entity.PageVO;
 import TwitchToYoutube.timeline.entity.Youtube;
 import TwitchToYoutube.timeline.manager.SessionManager;
-import TwitchToYoutube.timeline.manager.YoutubeService;
+import TwitchToYoutube.timeline.service.CommonService;
+import TwitchToYoutube.timeline.service.YoutubeService;
 import TwitchToYoutube.timeline.repository.YoutubeRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -11,11 +12,8 @@ import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.*;
 
+import javax.servlet.http.Cookie;
 import javax.servlet.http.HttpServletRequest;
-import java.net.MalformedURLException;
-import java.net.URL;
-import java.text.SimpleDateFormat;
-import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 
@@ -32,15 +30,11 @@ public class YoutubeController {
     @Autowired
     private YoutubeService service;
 
+    private CommonService common;
 
-    @PostMapping("/insert")
-    public String insertYoutube(HttpServletRequest request){
-        String referer = request.getHeader("Referer");
-        service.insertService(request);
-        return "redirect:/manage/1";
-    }
     @GetMapping("/{pageNum}")
     public String managePage(Model model,@PathVariable("pageNum") int pageNum){
+
         PageVO vo = service.getPageVO(pageNum);
         List<Youtube> list = service.getYoutubeList(vo);
         model.addAttribute("list",list);
@@ -51,17 +45,44 @@ public class YoutubeController {
     @ResponseBody
     @GetMapping("/find")
     public Map<String, String> findYoutube(@RequestParam String youtubeId){
+
         Map<String, String> map  = service.findYoutube(youtubeId);
         return map;
+    }
+    @PostMapping("/insert")
+    public String insertYoutube(HttpServletRequest request){
+        Map<String, String> user = (Map<String, String>) sessionManager.getSession(request);
+        Cookie streamer = common.findCookie(request, "streamer");
+        if(streamer.getValue().equals(user.get("login"))){
+            service.insertService(request);
+        }
+        String referer = request.getHeader("Referer");
+
+        return "redirect:/manage/1";
     }
 
     @PostMapping("/update")
     public String updateYoutube(HttpServletRequest request){
-        return service.updateYoutube(request);
+        Map<String, String> user = (Map<String, String>) sessionManager.getSession(request);
+        Cookie streamer = common.findCookie(request, "streamer");
+        if(streamer.getValue().equals(user.get("login"))){
+            return service.updateYoutube(request);
+        }
+        String referer = request.getHeader("Referer");
+
+        return "redirect:"+referer;
     }
 
     @PostMapping("/remove")
     public String removeYoutube(HttpServletRequest request){
-        return service.removeYoutube(request);
+        Map<String, String> user = (Map<String, String>) sessionManager.getSession(request);
+        Cookie streamer = common.findCookie(request, "streamer");
+        if(streamer.getValue().equals(user.get("login"))){
+            return service.removeYoutube(request);
+        }
+        String referer = request.getHeader("Referer");
+        return "redirect:"+referer;
     }
+
+
 }
